moved the tests to a folder and change the command for circleci
