+++ tests/stringifier.py
