+++ tests/test_bot.py
